==> mmp/rdkit/Kymera_ADMET/Archived/ADMETox_Update_1_2024Aug27/1_data_download.ipynb <==
import warnings
warnings.filterwarnings('ignore')

import os
import time

import pandas as pd

from d360api import d360api

##--------------------------------------------------------------
def dataDownload(my_query_id=2905, user_name="yjing@kymeratx.com", tokenFile='yjing_D360.token'):
    # Create API connection to the PROD server
    my_d360 = d360api(provider="https://10.3.20.47:8080")  # PROD environment
    user_name = user_name
    tokenFile = tokenFile
    
    with open(tokenFile, 'r') as ofh:
        service_token = ofh.readlines()[0]

    # Authenticate connection using service token
    my_d360.authenticate_servicetoken(servicetoken=service_token, user=user_name)
    results = my_d360.download_query_results(query_id=my_query_id)
    return results

##--------------------------------------------------------------
def loadCSV(dataTableFileName):
    try:
        dataTable = pd.read_csv(dataTableFileName)
    except Exception as e:
        print(f'Error: cannot read output file {dataTableFileName}; error msg: {e}')
        dataTable = None
    else:
        dataTable = dataTable.reset_index(drop=True)
        print(f'The downloaded data have data shape {dataTable.shape}')
    return dataTable

my_query_id = 3539 # 2905

## download data from D360 using API
dataTableFileName = dataDownload(my_query_id=my_query_id)
print(f'All data have been downloaded in file {dataTableFileName}')

## extract data from downloaded file
dataTable = loadCSV(dataTableFileName)
dataTable.head(3)
